api-rest/flask/requirements.txt
Flask==2.2.2
GitPython==3.1.31
kubernetes==24.2.0
requests==2.25.1
